Add test to consistency-check the concrete classes
All of the actual core functionality was already being
tested by the randomized/fuzzed Hypothesis tests, but
those dynamically subclass the base classes as needed,
and thus don't exercise the top-level classes which
users are most likely to use.

From an implementation-knowing perspective, I know
that those classes are literally empty except for
the ``size`` and ``formats`` definitions, so the
only thing left to test is that I didn't make any
obvious dumb mistake in defining those. So we just
test that the formats and sizes are consistent,
and hope that covers the most likely human errors
like accidentally deleting or adding a character.

This might be a good usecase for actual hardcoded
inputs too, just one for each format for each
class. But I'm fairly confident for now that I got
it right and won't likely break it in the future.